requirements.txt
streamlit==1.39.0
tensorflow==2.15.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
pillow==10.4.0
h5py==3.10.0
